+++ src/index.js
@@ -9,12 +9,8 @@
 
 import { JSONLanguage } from "./languages/json-language.js";
 import { JSONSourceCode } from "./languages/json-source-code.js";
-import noDuplicateKeys from "./rules/no-duplicate-keys.js";
-import noEmptyKeys from "./rules/no-empty-keys.js";
-import noUnsafeValues from "./rules/no-unsafe-values.js";
-import noUnnormalizedKeys from "./rules/no-unnormalized-keys.js";
-import sortKeys from "./rules/sort-keys.js";
-import topLevelInterop from "./rules/top-level-interop.js";
+import recommendedRules from "./build/recommended-config.js";
+import rules from "./build/rules.js";
 
 //-----------------------------------------------------------------------------
 // Plugin
@@ -23,30 +19,18 @@ import topLevelInterop from "./rules/top-level-interop.js";
 const plugin = {
 	meta: {
 		name: "@eslint/json",
-		version: "0.12.0", // x-release-please-version
+		version: "0.13.2", // x-release-please-version
 	},
 	languages: {
 		json: new JSONLanguage({ mode: "json" }),
 		jsonc: new JSONLanguage({ mode: "jsonc" }),
 		json5: new JSONLanguage({ mode: "json5" }),
 	},
-	rules: {
-		"no-duplicate-keys": noDuplicateKeys,
-		"no-empty-keys": noEmptyKeys,
-		"no-unsafe-values": noUnsafeValues,
-		"no-unnormalized-keys": noUnnormalizedKeys,
-		"sort-keys": sortKeys,
-		"top-level-interop": topLevelInterop,
-	},
+	rules,
 	configs: {
 		recommended: {
 			plugins: {},
-			rules: /** @type {const} */ ({
-				"json/no-duplicate-keys": "error",
-				"json/no-empty-keys": "error",
-				"json/no-unsafe-values": "error",
-				"json/no-unnormalized-keys": "error",
-			}),
+			rules: recommendedRules,
 		},
 	},
 };