+++ tools/update-readme.js
@@ -28,7 +28,7 @@ const README_FILE_PATH = "./README.md";
 
 /**
  * Fetches the latest sponsors from the website.
- * @returns {Promise<string>}} Prerendered sponsors markdown.
+ * @returns {Promise<string>} Prerendered sponsors markdown.
  */
 async function fetchSponsorsMarkdown() {
 	return got(SPONSORS_URL).text();
@@ -44,7 +44,7 @@ const allSponsors = await fetchSponsorsMarkdown();
 const readme = readFileSync(README_FILE_PATH, "utf8");
 
 let newReadme = readme.replace(
-	/<!--sponsorsstart-->[\w\W]*?<!--sponsorsend-->/u,
+	/<!--sponsorsstart-->[\s\S]*?<!--sponsorsend-->/u,
 	`<!--sponsorsstart-->\n\n${allSponsors}\n<!--sponsorsend-->`,
 );